+++ package.json
@@ -33,6 +33,7 @@
     "rehype-raw": "^7.0.0",
     "remark-gfm": "^4.0.0",
     "space-separated-tokens": "^2.0.2",
+    "use-immer": "^0.11.0",
     "web-vitals": "^4.2.3"
   },
   "scripts": {
